# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend",
    "description": "12-factor Rust backend: layered store + RBAC + JWT + WS + worker",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/system": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/system` — system monitoring dashboard data.",
        "operationId": "system_status",
        "responses": {
          "200": {
            "description": "System status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "`POST /api/auth/refresh` — rotate refresh token, issue new access.",
        "description": "Reads the refresh token from the httpOnly cookie first (preferred —\nthe token never touches JS), then falls back to the request body\n(for non-browser clients).",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rotated tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "required": [
          "user",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUser"
          }
        }
      },
...
      "DatabaseStats": {
        "type": "object",
        "required": [
          "backend",
          "urlMasked",
          "maxConnections",
          "minConnections",
          "activeConnections",
          "idleConnections",
          "sizeMb"
        ],
        "properties": {
          "activeConnections": {
            "type": "integer",
            "format": "int32",
            "description": "Current active connections (in use). `-1` if unavailable\n(e.g. SQLite which doesn't have a pool)."
          },
          "backend": {
            "type": "string",
            "description": "Backend name — `\"sqlite\"` or `\"postgres\"`."
          },
          "idleConnections": {
            "type": "integer",
            "format": "int32",
            "description": "Current idle connections (in pool, available). `-1` if unavailable."
          },
          "maxConnections": {
            "type": "integer",
            "format": "int32",
            "description": "Configured max pool connections.",
            "minimum": 0
          },
          "minConnections": {
            "type": "integer",
            "format": "int32",
            "description": "Configured min pool connections.",
            "minimum": 0
          },
          "sizeMb": {
            "type": "number",
            "format": "double",
            "description": "Database file size in MB (SQLite only — WAL + main DB file).\n`0.0` for Postgres (no file)."
          },
          "urlMasked": {
            "type": "string",
            "description": "Masked DB URL (password hidden)."
          }
        }
      },
...
      "ProcessStats": {
        "type": "object",
        "required": [
          "pid",
          "memoryMb",
          "virtualMemoryMb",
          "cpuUsage",
          "cpuCount",
          "osName",
          "osVersion",
          "hostname"
        ],
        "properties": {
          "cpuCount": {
            "type": "integer",
            "description": "Number of logical CPU cores.",
            "minimum": 0
          },
          "cpuUsage": {
            "type": "number",
            "format": "float",
            "description": "CPU usage % (0.0–100.0). Sampled over the last refresh interval.\nFirst call after startup may return 0.0 (sysinfo needs two\nsamples to compute usage)."
          },
          "hostname": {
            "type": "string",
            "description": "Hostname of the machine."
          },
          "memoryMb": {
            "type": "number",
            "format": "double",
            "description": "RSS memory in MB (resident set size — actual physical RAM used)."
          },
          "osName": {
            "type": "string",
            "description": "OS name (e.g. \"Linux\", \"macOS\", \"Windows\")."
          },
          "osVersion": {
            "type": "string",
            "description": "OS version / kernel version."
          },
          "pid": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "virtualMemoryMb": {
            "type": "number",
            "format": "double",
            "description": "Virtual memory in MB (total addressable memory, includes\nmapped files + shared libraries). Mostly informational."
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional refresh token in the body. If absent, the token is read\nfrom the `refresh_token` httpOnly cookie. Accepting it from the\nbody is a fallback for non-browser clients (curl, mobile) that\ncan't use cookies."
          }
        }
      },
...
      "SessionUser": {
        "type": "object",
        "description": "Identity of an authenticated actor, derived from the `user` row.\n\n`actor_type` mirrors the `user.role` column and is one of:\n  - `\"user\"`      — customer (book trips, feedback, ticket status)\n  - `\"employee\"`  — support staff (bookings, tickets, chat, calls, promos)\n  - `\"admin\"`     — full-access superuser (the first registered account)\n\nThe distinction matters for the chat / call hubs (only staff — employees\nOR admins — may register as agents) and for nullclaw (the human-fallback\nthreshold counts online staff).",
        "required": [
          "id",
          "type",
          "role",
          "name"
        ],
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "brandId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "brandName": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "employeeRole": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "SystemStatusResponse": {
        "type": "object",
        "required": [
          "uptime",
          "websocket",
          "database",
          "process"
        ],
        "properties": {
          "database": {
            "$ref": "#/components/schemas/DatabaseStats"
          },
          "process": {
            "$ref": "#/components/schemas/ProcessStats"
          },
          "uptime": {
            "$ref": "#/components/schemas/SystemUptime"
          },
          "websocket": {
            "$ref": "#/components/schemas/WebsocketStats"
          }
        }
      },
      "SystemUptime": {
        "type": "object",
        "required": [
          "seconds",
          "human"
        ],
        "properties": {
          "human": {
            "type": "string"
          },
          "seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "WebsocketStats": {
        "type": "object",
        "required": [
          "connections",
          "maxConnections",
          "rooms",
          "idempotencyEntries",
          "onlineEmployeeBrands",
          "onlineEmployees",
          "distinctIps"
        ],
        "properties": {
          "connections": {
            "type": "integer",
            "description": "Total live WS connections (across all users + employees).",
            "minimum": 0
          },
          "distinctIps": {
            "type": "integer",
            "description": "Distinct client IPs (used for per-IP connection caps).",
            "minimum": 0
          },
          "idempotencyEntries": {
            "type": "integer",
            "description": "Idempotency cache entries (pending `clientMsgId` lookups).",
            "minimum": 0
          },
          "maxConnections": {
            "type": "integer",
            "description": "Configured max connections (from `WS_MAX_CONNECTIONS`).",
            "minimum": 0
          },
          "onlineEmployeeBrands": {
            "type": "integer",
            "description": "Number of distinct brands with online employees. The hub\nindexes online employees by brand key (`brandId → employeeId →\nset<socketId>`). This is the count of brand keys with ≥1\nonline employee.",
            "minimum": 0
          },
          "onlineEmployees": {
            "type": "integer",
            "description": "Total online employees (sum across all brands). Useful for\nthe admin dashboard — \"how many support staff are online?\".",
            "minimum": 0
          },
          "rooms": {
            "type": "integer",
            "description": "Number of chat rooms (one per open channel that has at least\none joined socket). Empty rooms are cleaned up by the hub GC.",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}